Modifiche ai tipi e aggiunta liste utenti
Modifiche ai tipi. Ora tutto è uint16_t eccetto per i file descriptor che sono uint32_t.
Aggiunta della lista degli utenti. Il polling avviene tra tutti gli utenti, non solo tra quelli attivi.

+++ common.h
@@ -40,7 +40,7 @@ struct Messaggio_utente {
     uint16_t porta_utente;
 
     // Payload (usato per ACK_CARD, CARD_DONE, e CREATE_CARD)
-    int32_t id_card;
+    uint16_t id_card;
     
     // Payload (usato per CREATE_CARD)
     enum Colonne colonna;
